+++ traqo/__init__.py
@@ -6,7 +6,7 @@ import os
 
 from traqo._version import __version__
 from traqo.decorator import trace
-from traqo.tracer import Span, Tracer, get_tracer
+from traqo.tracer import Span, Tracer, get_current_span, get_tracer
 
 _disabled: bool = os.environ.get("TRAQO_DISABLED", "").strip() in ("1", "true", "yes")
 
@@ -28,6 +28,7 @@ __all__ = [
     "Tracer",
     "trace",
     "get_tracer",
+    "get_current_span",
     "disable",
     "enable",
     "__version__",